test(starry-kernel): co-locate axtest coverage cases into source crates
