+++ components/ax-lazyinit/Cargo.toml
@@ -10,3 +10,10 @@ categories = ["no-std", "rust-patterns"]
 license = "Apache-2.0"
 
 [dependencies]
+axtest = { workspace = true, optional = true }
+
+[features]
+axtest = ["dep:axtest"]
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(axtest)'] }
