+++ components/axfs-ng-vfs/src/lib.rs
@@ -4,6 +4,10 @@ extern crate alloc;
 #[cfg(test)]
 extern crate std;
 
+#[cfg(all(axtest, feature = "axtest"))]
+/// Coverage tests for VFS helpers and node contracts.
+pub mod axtest;
+
 mod fs;
 mod mount;
 mod node;
