+++ components/axio/Cargo.toml
@@ -11,9 +11,14 @@ license = "Apache-2.0"
 
 [features]
 alloc = []
+axtest = ["dep:axtest", "alloc"]
 default = []
 
 [dependencies]
 ax-errno = { workspace = true }
+axtest = { workspace = true, optional = true }
 heapless = "0.9"
 memchr = { version = "2", default-features = false }
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(axtest)'] }
