+++ components/axio/src/lib.rs
@@ -10,6 +10,10 @@
 #[cfg(feature = "alloc")]
 extern crate alloc;
 
+#[cfg(all(axtest, feature = "axtest"))]
+/// Coverage tests for I/O traits and adapters.
+pub mod axtest;
+
 #[doc(no_inline)]
 pub use ax_errno::{AxError as Error, AxErrorKind as ErrorKind, AxResult as Result};
