+++ components/cpumask/Cargo.toml
@@ -10,4 +10,11 @@ categories = ["os", "no-std"]
 license = "Apache-2.0"
 
 [dependencies]
+axtest = { workspace = true, optional = true }
 bitmaps = { version = "3.2.1", default-features = false }
+
+[features]
+axtest = ["dep:axtest"]
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(axtest)'] }
