+++ components/kernutil/src/lib.rs
@@ -1,5 +1,12 @@
 #![no_std]
 
+#[cfg(all(axtest, feature = "axtest"))]
+extern crate alloc;
+
+#[cfg(all(axtest, feature = "axtest"))]
+/// Coverage tests for kernel utility helpers.
+pub mod axtest;
+
 pub mod address;
 pub mod id;
 pub mod memory;
