+++ components/rsext4/src/lib.rs
@@ -40,6 +40,9 @@ pub use file::{
 pub use metadata::{chmod, chown, set_flags, set_project, utimens};
 
 pub mod api;
+#[cfg(all(axtest, feature = "axtest"))]
+/// Coverage tests for ext4 data structures and helpers.
+pub mod axtest;
 pub mod bitmap;
 pub mod blockdev;
 pub mod blockgroup_description;
