+++ components/scope-local/src/lib.rs
@@ -5,6 +5,10 @@
 
 extern crate alloc;
 
+#[cfg(all(axtest, feature = "axtest"))]
+/// Coverage tests for scoped local storage.
+pub mod axtest;
+
 mod boxed;
 mod item;
 mod scope;
